requirements.txt
pillow>=6.2.1
opencv-python>=4.1.2.30
tqdm>=4.41.0
numpy>=1.18.0